Refactor D2D sources logic, implement Unrealize()
